requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
joblib==1.4.2
lxml==5.2.1
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
soupsieve==2.5
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.1
